64 bit support

+++ src/lib/determ_clock.h
@@ -97,10 +97,10 @@
      int determ_task_clock_is_lowest();
      void determ_task_clock_start();
      void determ_task_clock_start_no_notify();
-     void determ_task_clock_stop_with_id(uint32_t id);
+     void determ_task_clock_stop_with_id(size_t id);
      void determ_task_clock_stop();
      void determ_task_clock_stop_no_notify();
-     void determ_task_clock_stop_with_id_no_notify(uint32_t id);
+     void determ_task_clock_stop_with_id_no_notify(size_t id);
      void determ_task_clock_halt();
      void determ_task_clock_activate();
      int determ_task_clock_activate_other(int32_t id);
@@ -114,7 +114,7 @@
      void determ_task_clock_close();
      void determ_task_clock_reset();
      u_int32_t determ_task_get_id();
-     int64_t determ_task_clock_estimate_next_tx(uint32_t id);
+     int64_t determ_task_clock_estimate_next_tx(size_t id);
      u_int64_t determ_task_clock_get_last_tx_size();
      struct determ_task_clock_info determ_task_clock_get_info();
      void determ_task_set_scaling_factor(uint8_t whole, uint8_t fraction);
